fix(invalid token):Check for a valid token before clearing token

+++ client/components/RequireAuthRoute.jsx
@@ -12,10 +12,13 @@ import { connect } from 'react-redux';
  * @returns {component} route or spinner
  */
 const RequireAuthRoute = ({ component: Component, path: route, authenticated }) => {
-  const decodeToken = decodeJwt(window.localStorage.token);
-  const exp = decodeToken.exp < new Date().getTime() / 1000;
-  if (exp === true) {
-    window.localStorage.clear();
+  const token = window.localStorage.getItem('token');
+  if (token !== null) {
+    const decodeToken = decodeJwt(token);
+    const exp = decodeToken.exp < new Date().getTime() / 1000;
+    if (exp === true) {
+      window.localStorage.clear();
+    }
   }
 
   const access = authenticated.loggedIn || window.localStorage.token !== undefined;
